Shorten strings in climbing aid lang parser to satisfy python lint

+++ lang/string_extractor/parsers/climbing_aid.py
@@ -10,21 +10,22 @@ def parse_climbing_aid(json, origin):
     elif "id" in json:
         name = json["id"]
 
+    down = None
     if "down" in json:
         down = json["down"]
 
-        if "menu_text" in down:
-            write_text(down["menu_text"], origin,
-                       comment="Menu option text for climbing down with aid \"{}\"".format(name))
-        if "menu_cant" in down:
-            write_text(down["menu_cant"], origin,
-                       comment="Menu option text when we can't climb down with climbing aid \"{}\"".format(name))
-        if "confirm_text" in down:
-            write_text(down["confirm_text"], origin,
-                       comment="Comfirmation prompt when climbing down with aid \"{}\"".format(name))
-        if "msg_before" in down:
-            write_text(down["msg_before"], origin,
-                       comment="Message shown before climbing down with aid \"{}\"".format(name))
-        if "msg_after" in down:
-            write_text(down["msg_after"], origin,
-                       comment="Message shown after climbing down with aid \"{}\"".format(name))
+    if down and "menu_text" in down:
+        write_text(down["menu_text"], origin,
+                    comment="Menu text for descent with \"{}\"".format(name))
+    if down and "menu_cant" in down:
+        write_text(down["menu_cant"], origin,
+                    comment="Menu disables descent with \"{}\"".format(name))
+    if down and "confirm_text" in down:
+        write_text(down["confirm_text"], origin,
+                    comment="Query: really descend with \"{}\"?".format(name))
+    if down and "msg_before" in down:
+        write_text(down["msg_before"], origin,
+                    comment="Message before descent with \"{}\"".format(name))
+    if down and "msg_after" in down:
+        write_text(down["msg_after"], origin,
+                    comment="Message after descent with \"{}\"".format(name))
